// frontend/src/api/models/MazesMazeIdGet200ResponseNodesInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Maze API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MazesMazeIdGet200ResponseNodesInner
 */
export interface MazesMazeIdGet200ResponseNodesInner {
    /**
     * 
     * @type {number}
     * @memberof MazesMazeIdGet200ResponseNodesInner
     */
    mazeNodeId: number;
    /**
     * 
     * @type {number}
     * @memberof MazesMazeIdGet200ResponseNodesInner
     */
    x: number;
    /**
     * 
     * @type {number}
     * @memberof MazesMazeIdGet200ResponseNodesInner
     */
    y: number;
}

/**
 * Check if a given object implements the MazesMazeIdGet200ResponseNodesInner interface.
 */
export function instanceOfMazesMazeIdGet200ResponseNodesInner(value: object): value is MazesMazeIdGet200ResponseNodesInner {
    if (!('mazeNodeId' in value) || value['mazeNodeId'] === undefined) return false;
    if (!('x' in value) || value['x'] === undefined) return false;
    if (!('y' in value) || value['y'] === undefined) return false;
    return true;
}

export function MazesMazeIdGet200ResponseNodesInnerFromJSON(json: any): MazesMazeIdGet200ResponseNodesInner {
    return MazesMazeIdGet200ResponseNodesInnerFromJSONTyped(json, false);
}

export function MazesMazeIdGet200ResponseNodesInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): MazesMazeIdGet200ResponseNodesInner {
    if (json == null) {
        return json;
    }
    return {
        
        'mazeNodeId': json['mazeNodeId'],
        'x': json['x'],
        'y': json['y'],
    };
}

export function MazesMazeIdGet200ResponseNodesInnerToJSON(json: any): MazesMazeIdGet200ResponseNodesInner {
    return MazesMazeIdGet200ResponseNodesInnerToJSONTyped(json, false);
}

export function MazesMazeIdGet200ResponseNodesInnerToJSONTyped(value?: MazesMazeIdGet200ResponseNodesInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'mazeNodeId': value['mazeNodeId'],
        'x': value['x'],
        'y': value['y'],
    };
}